requirements.txt
python-dotenv
requests
urllib3
redis
websocket-client
aiohttp